refactor(tests): remove fake simulation logic + break cyclic imports
- fix cyclic imports by using nested sub-barrel imports (/element, /structure, etc.) instead of main  barrel within src/lib/
- move toggle_fullscreen and InfoItem to /layout/fullscreen.ts
- move is_binary to /io/is-binary.ts
- move FileInfo to /io/types.ts
