+++ src/lib/convex-hull/barycentric-coords.ts
@@ -1,4 +1,4 @@
-import type { ElementSymbol } from '$lib'
+import type { ElementSymbol } from '$lib/element'
 import type { Vec3 } from '$lib/math'
 import { compute_e_form_per_atom, find_lowest_energy_unary_refs } from './thermodynamics'
 import type { ConvexHullEntry, PhaseData, Point3D } from './types'
