+++ src/lib/io/types.ts
@@ -1,0 +1,7 @@
+export interface FileInfo {
+  name: string
+  url: string
+  type?: string
+  category?: string
+  category_icon?: string
+}
