+++ src/lib/structure/StructureExportPane.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-  import type { AnyStructure } from '$lib'
+  import type { AnyStructure } from '$lib/structure'
   import DraggablePane from '$lib/overlays/DraggablePane.svelte'
   import { export_canvas_as_png } from '$lib/io/export'
   import * as exports from '$lib/structure/export'
